% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proportion_overlap}
\alias{proportion_overlap}
\title{Calculate proportion overlapping of intersecting intervals}
\usage{
proportion_overlap(...)
}
\arguments{
\item{...}{Interval matrices (passed to \code{\link{join}}).}
}
\value{
\code{data.frame} containing integer columns corresponding to indices of intervals within the input matrices and a final numeric column called \code{proportion_overlap} containing the fraction of the size of the intersection within the union.
}
\description{
Proportion overlapping is calculated as the size of the intersection of intervals, divided by the size of the union.
}
\examples{
proportion_overlap(rbind(c(1, 2), c(49, 51), c(50, 200)), rbind(c(50, 100)))
}
