% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sections}
\alias{sections}
\title{Get the sections from a set of interval breaks}
\usage{
sections(x)
}
\arguments{
\item{x}{Sorted integer vector.}
}
\value{
Intervals represented by integer matrix of two columns.
}
\description{
Given a set of interval breaks (see \code{\link{breaks}}), generate a new set of intervals, the `sections', which partitions the full range of the given set, with an interval between every `break' (i.e. start/end point) in the given set.
}
\examples{
sections(1:10)
}
